Simplify the `.closest(Array)` code since we no longer use it internally.
It was previously used in liveHandler for delegated events, but is now baked in to events.js. The signature is a strange one since it returns an array of matches, *not* a chainable jQuery object. I've marked it as deprecated.

The code is simplified by two things: 1) We don't care about optimizing the duplicate selectors case, and 2) `.is()` now handles positional selectors by looking for set membership, which eliminates the workaround that was being applied here. See https://github.com/jquery/jquery/commit/70e2e32e0eb03607ad0c8b7752dbd7747da47164 for that commit.